requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.42.0
blinker==1.8.2
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.23.0
netifaces==0.11.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.0.3
pillow==10.4.0
ping3==4.0.8
pkgutil-resolve-name==1.3.10
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==17.0.0
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
safetensors==0.5.2
scipy==1.10.1
six==1.17.0
smmap==5.0.2
speedtest-cli==2.1.3
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.41.0
triton==3.0.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
watchdog==4.0.2
werkzeug==3.0.6
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
